+++ src/main/java/shop/itbook/itbookshop/productgroup/product/dto/response/ProductDetailsResponseDto.java
@@ -39,7 +39,7 @@ public class ProductDetailsResponseDto {
 
     private Long fixedPrice;
 
-    private Double discountPercent;
+    private Integer discountPercent;
 
     private String fileThumbnailsUrl;
 
@@ -77,7 +77,7 @@ public class ProductDetailsResponseDto {
                                      String detailsDescription, Boolean isSelled,
                                      Boolean isForceSoldOut,
                                      Integer stock, Integer increasePointPercent, Long rawPrice,
-                                     Long fixedPrice, Double discountPercent,
+                                     Long fixedPrice, Integer discountPercent,
                                      String fileThumbnailsUrl,
                                      String isbn, Integer pageCount, LocalDateTime bookCreatedAt,
                                      Boolean isEbook, String fileEbookUrl, String publisherName,
