+++ src/main/java/shop/itbook/itbookshop/productgroup/product/entity/SearchProduct.java
@@ -58,7 +58,7 @@ public class SearchProduct {
     private Integer increasePointPercent;
 
     @Column
-    private Double discountPercent;
+    private Integer discountPercent;
 
     @Column
     private Long rawPrice;
@@ -82,9 +82,10 @@ public class SearchProduct {
     @SuppressWarnings("java:S107") // 생성자 필드 갯수가 많아 추가
     @Builder
     public SearchProduct(Long productNo, String name, String simpleDescription,
-                         String detailsDescription, Integer stock, Boolean isSelled, Boolean isForceSoldOut,
+                         String detailsDescription, Integer stock, Boolean isSelled,
+                         Boolean isForceSoldOut,
                          String thumbnailUrl, Long fixedPrice,
-                         Integer increasePointPercent, Double discountPercent, Long rawPrice) {
+                         Integer increasePointPercent, Integer discountPercent, Long rawPrice) {
         this.productNo = productNo;
         this.name = name;
         this.simpleDescription = simpleDescription;
